refactor: split email-suppression functions into their file email-suppression.ts to make mocking in sendEmailGroups more convenient, and to further match structure of mocks/email-suppression.ts
